requirements.txt
fastapi==0.103.2
pydantic==2.4.2
sqlalchemy==2.0.22
uvicorn==0.23.2
python-dotenv==1.0.0
email-validator==2.0.0
